First upload
Uploaded the xArm6 files:
- STLs
- DAEs
- Solidwork Part files

Also, I uploaded the blender file with all the joints defined as bones and simple inverse kinematics applied to the system for the motion demo.

A folder that containing all the animation images of a scanning task is uploaded for reference.

The animation is generated using EEVEE to save time and computational power.