Removing animation from manipulator base files
For better organisation, the base blender files for the manipulators will no have any animation in its keyframes.

This allow users to just append directly without the need to keep removing the animation if there is any.

With that in mind, any manipulator with animations that is for demostration purposes will be stored under the "examples" folder.